move tests back into _test_multiprocessing and use @classmethod in order to make the worker object accessible/serializable on Windows
